Create ecclean_macos
This script provides the ability to easily remove synced blockchain files (located in ~/Library/Application Support/eccoin) from previous fork. Failure to remove these files results in the inability of the current daemon to properly connect to the network and sync.  This behavior is observed with clients that have synced with daemon version < 0.2.5.10.  
In addition to removal of the data files located in (~/Library/Application Support/eccoin), This script also allows for easy deletion of the daemon and version file from (~/Library/Application Support/.eccoin-wallet) or to simply check the installed version, if it exists.  Deletion of the current daemon will force sapphire to download the latest available daemon release upon next start.